# doc/api/swagger.yaml
basePath: /api/v2
definitions:
...
  database.Signal:
    properties:
      direction:
        description: Direction of the signal (in or out)
        type: string
      id:
        type: integer
      index:
        description: Index of the Signal in the mapping
        type: integer
      name:
        description: Name of Signal
        type: string
      simulationModelID:
        description: ID of simulation model
        type: integer
      unit:
        description: Unit of Signal
        type: string
    type: object
...
  docs.ResponseError:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  docs.ResponseSignal:
    properties:
      signal:
        $ref: '#/definitions/database.Signal'
        type: object
    type: object
  docs.ResponseSignals:
    properties:
      signals:
        items:
          $ref: '#/definitions/database.Signal'
        type: array
    type: object
...
  signal.addSignalRequest:
    properties:
      signal:
        $ref: '#/definitions/signal.validNewSignal'
        type: object
    type: object
  signal.updateSignalRequest:
    properties:
      signal:
        $ref: '#/definitions/signal.validUpdatedSignal'
        type: object
    type: object
  signal.validNewSignal:
    properties:
      Direction:
        type: string
      Index:
        type: integer
      Name:
        type: string
      SimulationModelID:
        type: integer
      Unit:
        type: string
    required:
    - Direction
    - Index
    - Name
    - SimulationModelID
    type: object
  signal.validUpdatedSignal:
    properties:
      Index:
        type: integer
      Name:
        type: string
      Unit:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: sonja.happ@eonerc.rwth-aachen.de
    name: Sonja Happ
  description: |-
    This is the VILLASweb Backend API v2.0.
    Parts of this API are still in development. Please check https://git.rwth-aachen.de/acs/public/villas/web-backend-go for more information.
  license:
    name: GNU GPL 3.0
    url: http://www.gnu.de/documents/gpl-3.0.en.html
  title: VILLASweb Backend API
  version: "2.0"
paths:
...
  /signals:
    get:
      operationId: getSignals
      parameters:
      - description: Direction of signal (in or out)
        in: query
        name: direction
        required: true
        type: string
      - description: Model ID of signals to be obtained
        in: query
        name: modelID
        required: true
        type: string
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Signals which belong to simulation model
          schema:
            $ref: '#/definitions/docs.ResponseSignals'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get all signals of one direction
      tags:
      - signals
    post:
      consumes:
      - application/json
      operationId: AddSignal
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: A signal to be added to the model incl. direction and model ID
          to which signal shall be added
        in: body
        name: inputSignal
        required: true
        schema:
          $ref: '#/definitions/signal.addSignalRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Signal that was added
          schema:
            $ref: '#/definitions/docs.ResponseSignal'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Add a signal to a signal mapping of a model
      tags:
      - signals
  /signals/{signalID}:
    delete:
      operationId: deleteSignal
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of signal to be deleted
        in: path
        name: signalID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Signal that was deleted
          schema:
            $ref: '#/definitions/docs.ResponseSignal'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Delete a signal
      tags:
      - signals
    get:
      operationId: getSignal
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of signal to be obtained
        in: path
        name: signalID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Signal that was requested
          schema:
            $ref: '#/definitions/docs.ResponseSignal'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get a signal
      tags:
      - signals
    put:
      operationId: updateSignal
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: A signal to be updated
        in: body
        name: inputSignal
        required: true
        schema:
          $ref: '#/definitions/signal.updateSignalRequest'
          type: object
      - description: ID of signal to be updated
        in: path
        name: signalID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Signal that was updated
          schema:
            $ref: '#/definitions/docs.ResponseSignal'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Update a signal
      tags:
      - signals
...
swagger: "2.0"